services/kafka_bridge/requirements.txt
kafka-python>=2.0
psycopg2-binary>=2.9
python-dateutil>=2.8
